Fix bug causing the module to freeze
This bug can be reproduced in the previous build by using Penrose in
continuous mode (no trigger input) with a rapidly changing CV input
such as a sloped waveform from a VCO, and with all notes enabled. This
causes the quantizer to write to the DAC over SPI extremely frequently.
Then, start turning on and off the C, C#, D, and D# notes. Eventually
the module will lock up.

Those four buttons are connected to row 1 of the button matrix, which
is pin PB2. In the previous build, the matrix rows were inputs and the
columns were outputs, but PB2 is also the SS pin of the SPI peripheral.
According to the ATmega168 datasheet (DS40002074A section 19.3 'SS Pin
functionality'):

  If SS is configured as an input, it must be held high to ensure Master
  SPI operation. If the SS pin is driven low by peripheral circuitry
  when the SPI is configured as a Master with the SS pin defined as an
  input, the SPI system interprets this as another master selecting
  the SPI as a slave and starting to send data to it. To avoid bus
  contention, the SPI system takes the following actions:

  1. The MSTR bit in SPCR is cleared and the SPI system becomes a Slave.
     As a result of the SPI becoming a Slave, the MOSI and SCK pins
     become inputs.

The bug happens when, by chance, the user presses one of those four
buttons while row 1 and the button's column in the matrix are selected
for polling and while the SPI peripheral is transmitting. Since the
selected column is driven low, the SS pin will be driven low as well,
thereby disrupting the SPI transmission and causing the SPI peripheral
to become a slave. This happens despite the fact that the DAC output
routine sets the column outputs high just before transmitting, probably
due to parasitics. Then, since the firmware busy-waits for SPI
transmission completion, any further attempts to transmit will enter an
infinite loop, since there is no master on the bus.

The solution is to 'invert' the button matrix by making the rows
outputs and the columns inputs, since the SS pin has no effect on the
SPI peripheral when it is configured as an output.

The functions io_processButtons() and io_isButtonPushed() in IoMatrix.c
would also require 'inversion' but they were completely unused so I've
just removed them instead.

An alternative solution is to simply set the MSTR bit in SPCR after
every SPI transmission, but the solution presented here prevents the
SPI from ever becoming a slave in the first place.